refactor(hypervisor): split linux vm
